requirements.txt
fastapi==0.116.1
uvicorn==0.35.1
websockets-15.0.1
openai==1.100.2
python-multipart==0.0.20
python-dotenv==1.1.1
